+++ src/Bucket.ts
@@ -9,6 +9,7 @@ import { Readable } from "stream";
 import { Buffer } from "buffer";
 import { Batch, BatchType } from "./Batch";
 import { isCompatibale } from "./Client";
+import Stream from "stream";
 
 /**
  * Options for querying records
@@ -40,6 +41,7 @@ export interface WriteOptions {
 export class Bucket {
   readonly name: string;
   private httpClient: AxiosInstance;
+  private isBrowser: boolean;
 
   /**
    * Create a bucket. Use Client.creatBucket or Client.getBucket instead it
@@ -51,6 +53,7 @@ export class Bucket {
   constructor(name: string, httpClient: AxiosInstance) {
     this.name = name;
     this.httpClient = httpClient;
+    this.isBrowser = typeof window !== "undefined";
     this.readRecord = this.readRecord.bind(this);
   }
 
@@ -283,7 +286,7 @@ export class Bucket {
     const { data, headers } = await this.httpClient.get(url);
     const { id } = data;
     const header_api_version = headers["x-reduct-api"];
-    if (isCompatibale("1.5", header_api_version)) {
+    if (isCompatibale("1.5", header_api_version) && !this.isBrowser) {
       yield* this.fetchAndParseBatchedRecords(
         entry,
         id,
@@ -345,7 +348,11 @@ export class Bucket {
     const request = head ? this.httpClient.head : this.httpClient.get;
     const { status, headers, data } = await request(
       `/b/${this.name}/${entry}?${param}`,
-      head ? undefined : { responseType: "stream" },
+      head
+        ? undefined
+        : {
+            responseType: this.isBrowser ? "arraybuffer" : "stream",
+          },
     );
 
     if (status === 204) {
@@ -361,14 +368,32 @@ export class Bucket {
       }
     }
 
-    return new ReadableRecord(
-      BigInt(headers["x-reduct-time"] ?? 0),
-      BigInt(headers["content-length"] ?? 0),
-      headers["x-reduct-last"] == "1",
-      data,
-      labels,
-      headers["content-type"] ?? "application/octet-stream",
-    );
+    if (this.isBrowser) {
+      // Pass a dummy Stream object and use ArrayBuffer
+      const arrayBuffer = data as ArrayBuffer;
+      return new ReadableRecord(
+        BigInt(headers["x-reduct-time"] ?? 0),
+        BigInt(headers["content-length"] ?? 0),
+        headers["x-reduct-last"] == "1",
+        head,
+        new Stream.Readable(),
+        labels,
+        headers["content-type"] ?? "application/octet-stream",
+        arrayBuffer,
+      );
+    } else {
+      // Pass the actual Stream object to ReadableRecord
+      const stream = data as Readable;
+      return new ReadableRecord(
+        BigInt(headers["x-reduct-time"] ?? 0),
+        BigInt(headers["content-length"] ?? 0),
+        headers["x-reduct-last"] == "1",
+        head,
+        stream,
+        labels,
+        headers["content-type"] ?? "application/octet-stream",
+      );
+    }
   }
 
   private async *fetchAndParseBatchedRecords(
@@ -464,6 +489,7 @@ export class Bucket {
         BigInt(ts),
         size,
         last,
+        head,
         stream,
         labels,
         contentType,