+++ src/Record.ts
@@ -1,4 +1,5 @@
 import Stream from "stream";
+import { Buffer } from "buffer";
 // @ts-ignore`
 import { AxiosInstance } from "axios";
 import { WriteOptions } from "./Bucket";
@@ -15,6 +16,7 @@ export class ReadableRecord {
   public readonly stream: Stream;
   public readonly labels: LabelMap = {};
   public readonly contentType: string | undefined;
+  private readonly arrayBuffer: ArrayBuffer | undefined;
 
   /**
    * Constructor which should be call from Bucket
@@ -24,9 +26,11 @@ export class ReadableRecord {
     time: bigint,
     size: bigint,
     last: boolean,
+    head: boolean,
     stream: Stream,
     labels: LabelMap,
     contentType?: string,
+    arrayBuffer?: ArrayBuffer,
   ) {
     this.time = time;
     this.size = size;
@@ -34,17 +38,31 @@ export class ReadableRecord {
     this.stream = stream;
     this.labels = labels;
     this.contentType = contentType;
+    this.arrayBuffer = arrayBuffer;
+
+    if (head) {
+      (this.arrayBuffer as ArrayBuffer) = new ArrayBuffer(0);
+    }
   }
 
   /**
    * Read content of record
    */
   public async read(): Promise<Buffer> {
+    if (this.arrayBuffer !== undefined) {
+      return new Promise((resolve, reject) => {
+        try {
+          resolve(Buffer.from(this.arrayBuffer as ArrayBuffer));
+        } catch (error) {
+          reject(error);
+        }
+      });
+    }
     const chunks: Buffer[] = [];
     return new Promise((resolve, reject) => {
-      this.stream.on("data", (chunk: Buffer) => chunks.push(chunk));
-      this.stream.on("error", (err: Error) => reject(err));
-      this.stream.on("end", () => resolve(Buffer.concat(chunks)));
+      (this.stream as Stream).on("data", (chunk: Buffer) => chunks.push(chunk));
+      (this.stream as Stream).on("error", (err: Error) => reject(err));
+      (this.stream as Stream).on("end", () => resolve(Buffer.concat(chunks)));
     });
   }