+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added:
 
 - RS-31: `Bucket.update` and `Bucket.beginUpdateBatch` methods for changing labels, [PR-89](https://github.com/reductstore/reduct-js/pull/89)
+- RS-394: Add query support for browser, [PR-90](https://github.com/reductstore/reduct-js/pull/90)
 
 ## [1.10.1] - 2024-07-01